backend_django/requirements.txt
Django==4.2.10
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
python-dotenv>=1.0.0
drf-yasg>=1.21.7
mongoengine>=0.27.0
pymongo==3.12.3
dnspython>=2.3.0
sqlparse>=0.3.1 